+++ notification/src/epilot/models/operations/gettotalunread.py
@@ -12,9 +12,9 @@ class GetTotalUnreadResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    get_total_unread_200_text_plain_number: Optional[str] = dataclasses.field(default=None)
-    r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
+    res: Optional[str] = dataclasses.field(default=None)
+    r"""Success"""