+++ notification/src/epilot/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.3.0'
-    gen_version: str = '2.173.0'
-    user_agent: str = 'speakeasy-sdk/python 1.3.0 2.173.0 1.0.0 epilot-notification'
+    sdk_version: str = '2.0.0'
+    gen_version: str = '2.181.1'
+    user_agent: str = 'speakeasy-sdk/python 2.0.0 2.181.1 1.0.0 epilot-notification'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: